Use `#[repr)C)]` for TraitObject to ensure correct layout
Rust doc states that:
There is no indirection for these types; all data is stored within the
struct, as you would expect in C. However with the exception of arrays
(which are densely packed and in-order), the layout of data is not
specified by default.

We can't assume layout of `#[repr(rust)] struct Type`, so use
`#[repc(C)] TraitObject` to ensure correct memory layout.

Signed-off-by: Liu Jiang <gerry@linux.alibaba.com>